integral now int not uint

+++ cass_pnccd/pnccd_analysis.h
@@ -105,7 +105,7 @@ namespace cass
       bool            _createPixellist;   //!< flag to switch pixellist on / off
       bool            _doOffsetCorrection;//!< flag to switch offsetcorrection on / off
       bool            _useCommonMode;     //!< flag to switch a common mode subtraction scheme
-      uint32_t        _thres_for_integral;//!< the thresold for special integral
+      int64_t        _thres_for_integral;//!< the thresold for special integral
       std::string     _darkcalfilename;   //!< filename of file containing dark & noisemap
       std::string     _savedarkcalfilename;// Dark frame calibration save file names for each detector, it is automatically generated//
       cass::detROI_   _detROI;
